update circuit type
